Cleaning up.
git-svn-id: svn+ssh://cv077/home/ueshiba/svnroot/trunk/src@2062 3ef1b8b6-fb57-41bf-88d9-9a5bd220bf38

+++ TUXv++/TU/v/XvDC.h
@@ -44,7 +44,7 @@ class XvDC : public ShmDC, public List<XvDC>::Node
 {
   public:
     XvDC(CanvasPane& parentCanvasPane,
-	 u_int width=0, u_int height=0, float zoom=1)			;
+	 size_t width=0, size_t height=0, float zoom=1)			;
     virtual		~XvDC()						;
 
     using		ShmDC::operator <<;
